feat(save): Massive revision of the Save type
Instead of a sticking all countries, competitions, clubs, players and
matches in a single, large, nested object called allCompetitions,
each of those objects has been broken out into it's own object and
added directly to a save. Each object now has "references" to objects
at a lower, e.g. each club has the ID and names of each of it's
players instead of the full player object. Besides just making it
signficantly easier to work on advancing the game state, this makes it
significantly easier to promote and relegate clubs, buy and sell
players as well as organize international cups like the Champions
League. This will also significantly improve the speed and
reponsiveness of this game.